requirements.txt
cohere>=5.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn>=0.29.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
